Schemes no longer part of the installed package
We opt to only package and install the framework portions of charm
instead of the schemes. Installings the schemes makes it  needlessly
difficutl to work withthem, test them, etc. If we did install them a
long with everything else, there would be an odd distinction between
the schemes that were last installed and anything a cryptographer has
written recently or worse modified.