Scheme tests moved to schemes/test/ and rm'd tests/
Schemes should be with their own tests. This follows form the fact
that we want to ship the core tests with the charm framework.
As such, we to have them in the charm package. This leaves
tests for schemes orphined. Moreover, we should ship the scheme tests
when we ship schemes (whether it is as part of a package or as a
downloadable package of examples). Thus they are now in schemes/test