Renaming charm to charm-pkg and fixing setup.py
Because we split schemes of, we epect people to run the python
intepreter from the project directory. As, because of python search
order for imports, the intepreter would actually try to import
project/charm instead of installed version. This would break because
the project version does not have the binaries built for it.