Reverts/Corrects fix for benchmark.so import on OSX
Restructuring charm caused an old OSX bug to reappear:
any import of a extension module that used
benchmark to fail with "No module named benchmark."
Ayo re-introduced the old __init__.py for charm in order
to fix what was proceved as a pathing issue with finding benchark.so.

In fact, the issue is not pathing but dynamic linking and symbol
resolution:in order for symbols to resolve benhmark.so must be
loaded first or symbol resolution when importing any dependant module
(e.g. integer.so,pairing.so,etc) can fail. It appears that symbol
resolution happens when the so (e.g. integer.so) is loaded, but that
some python interpreters/operating systems don't import the
libraries it's linked against untill we explicity call the
import_benchmark function. In this case, symbol resolution fails.
This happens on OSX but not apparently on ubuntu.

This reverts commits 90d22ea8fa8e134785a2165fe036b78f932b5b3d.
and 985d3e48c7fe134c925255cf8214e5b138623fff which had a
pathing hack that incidently fixed the issue by importing
benchmark and replaces them with an explicit import call
when the charm package is loaded via __init__.py