Adding Distribute to handle pyparsing dependency
This commit fixes #33 by using the python distribute package
to handle dependecy resolution for python packages charm depend on.
It adds the bootstraping script for installing distribute, switches
setup.py to pull setup(...) from  setuptools(this is the recommended
practice since  distribute is a drop in replacement setuptools), adds
the dependency on pyparsing 1.5.5, and finally removes the old
pyparsing install logic.

Pior to this, the old download logic in ./configure and the
Makefile only checked for the existence of pyparsing and not the
version, causing failures when an old version is installed.

The minimal fix would be to add version logic. However,were we to do
so we would be going further and further into handling dependency
managment. That path leads to darkness and dispair.

Distribute, a drop in replacement for setup tools, handles dependecy
resolution issues itself. Further, it allows for bootstraping itself
via the now included distribute_setup.py script. Finally, by making
more of the dependencies known to setup.py we make it far easier to make
binary distributions such as windows/osx installaers and debain pacakges.