Configure.sh searches for python,python3,python3.2 etc
Fenc on osx seems to be default install python3.2 as python3.2
not python or python 3.  ./Configure now either trys the argument
from --python and prints and error message and  exits
if the supplied argument is  not python3.If no arg is given it trys
python, then python3, then python3.2 then python3.1 before outputing
and error message and exiting if none of those work.