Renames charm-framwork to charm and fixes imports
python setup.py test, which is the command we use to run tests now.
first runs a build, the temporarly adds the current directory to test,
and then runs the test pass (via pytest in our case). The idea is
that the tests will be run against the code in the directory, not
installed tests.

However, because we renamed charm to charm-framework, nothing is imported
and then the installed version of charm is used instead. This causes
strange test results.

We change charm-framwork back to charm/ to enable the expected
behavior.

Also, this fixes an import in the docttest for the waters scheme