Making extensions build in place by default
By specifying build_ext inplace in the config file,any build command
will end up copying the built .so modules into the charm pacakge.
This prevents people from installing charm and then running into errors
where python imports the local copy that does not have the compiled
c extensions. Although this could cause issues because people might
believe they are running with the installed copy and not local version,
such issues should only matter if the local version was modified.
At the point that someone is doing that, they are changing the
core components of charm, and it seems fair to assume they either
understand python import resulution or need to.

It is worth noting that the disutils documentation explicitly states
that this is not a good idea  "since always building extensions in-place
would break installation of the module distribution." Well this may have
been true for distuils, it appears not to be true for distribute. All
the option appears to do is insure that after anything is build, the
shared objects are coppied over. As such it doesn't break anything.
setup.py install still works.