Adds setup.cfg to manifest
This appears to fix #41 in its entirity.  By adding setup.cfg, we ensure
everything is built inplace and hence the version of charm thats imported
works. This fixes the benchmark import issue.

It's not clear why this fixes the issue we saw with pyparsing, but we
can't repro it again after the fix.