Adding defualts for config values in setup.py
This fixes #39 by allowing us not to include config.mk
in the source distribution but still have tox function.

In general, the full solution would be to remove ./configure in
its entirity since it is only used to handle dependency installation.
Realistically, we should follow standard practice on installing from
source and not handle dependencies. It is up to the user to install
openssl-dev, gmp, and ppc. These dependencies, however, should be
handeled by installers.