Added support for uncompressed curves elements (de)serialization.
Serializing curve elements through their compressed representation is sometime
not convenient due to the large computational costs of the "decompression"
algorithm of some curve.

The serialize and deserialize routines of the c extension 'pairing' now take
an additional positional argument to disable elements compression.

The corresponding serialize and deserialize methods of the PairingGroup class
now implements an optional keyword-only argument 'compression'.

The default value behavior is to use compression for compatibility with
previous versions.