symcrypto.py: update to allow for AEAD
Added the AEAD capability to symcrypto.py, i.e., it is possible to (optionally) specify additional data to be MACed together with the ciphertext. The additional data is typically a plaintext that will not be encrypted, but it is desired that this additional data be authenticated with the ciphertext. For instance, a header can be added to the ciphertext to be authenticated. Example of usage is included in the class docstring.

The module is backwards compatible with the previous AE (Authenticated Encryption) version with no additional data. All adapter classes that utilize symcrypto without additional data should continue working with no modifications.