Add a 254-bit Type-F curve to PBC
These parameters for a BN254 curve were generated using the genfparams program from the PBC library. Without this change, it is not possible to use the BN254 curve without using the optional Relic library. This change does not appear to degrade existing tests; however, it may be useful to create some tests that directly test the functionality of the Type-F curve. I have manually tested this by using it in a new pairing-based traitor tracing implementation.