Update setup.py
In [Pyparsing 2.4.1](https://github.com/pyparsing/pyparsing/releases/tag/pyparsing_2.4.1), `downcaseTokens` is moved to `pyparsing_common`, breaking imports in policytree.py. 

An alternative fix is to update policytree.py to use the `pyparsing_common.downcaseTokens` import, but that might involve a permanent bump to the pyparsing version (which is the correct fix, but less conservative).